tenantcostclient: relocate to pkg/multitenant/tenantcostclient
The package has no CCL dependencies, so its home under
pkg/ccl/multitenantccl was incidental. Move it next to the rest of
pkg/multitenant.

pkg/ccl/multitenantccl/doc.go keeps a blank import of the new location
so callers still pull in the package's init hooks. That indirection
goes away in a later commit.

Epic: none
Release note: None
