+++ pkg/BUILD.bazel
@@ -41,7 +41,6 @@ ALL_TESTS = [
     "//pkg/ccl/changefeedccl:changefeedccl_test",
     "//pkg/ccl/multiregionccl:multiregionccl_test",
     "//pkg/ccl/multitenantccl/tenantcapabilitiesccl:tenantcapabilitiesccl_test",
-    "//pkg/ccl/multitenantccl/tenantcostclient:tenantcostclient_test",
     "//pkg/ccl/multitenantccl/tenantcostserver/tenanttokenbucket:tenanttokenbucket_test",
     "//pkg/ccl/multitenantccl/tenantcostserver:tenantcostserver_test",
     "//pkg/ccl/schemachangerccl/sctestbackupccl:sctestbackupccl_test",
@@ -301,6 +300,7 @@ ALL_TESTS = [
     "//pkg/multitenant/tenantcapabilities/tenantcapabilitiesauthorizer:tenantcapabilitiesauthorizer_test",
     "//pkg/multitenant/tenantcapabilities/tenantcapabilitieswatcher:tenantcapabilitieswatcher_test",
     "//pkg/multitenant/tenantcapabilities:tenantcapabilities_test",
+    "//pkg/multitenant/tenantcostclient:tenantcostclient_test",
     "//pkg/multitenant/tenantcostmodel:tenantcostmodel_test",
     "//pkg/obs/ash:ash_test",
     "//pkg/obs/clustermetrics/cmreader:cmreader_test",
@@ -1022,8 +1022,6 @@ GO_TARGETS = [
     "//pkg/ccl/multiregionccl:multiregionccl",
     "//pkg/ccl/multiregionccl:multiregionccl_test",
     "//pkg/ccl/multitenantccl/tenantcapabilitiesccl:tenantcapabilitiesccl_test",
-    "//pkg/ccl/multitenantccl/tenantcostclient:tenantcostclient",
-    "//pkg/ccl/multitenantccl/tenantcostclient:tenantcostclient_test",
     "//pkg/ccl/multitenantccl/tenantcostserver/tenanttokenbucket:tenanttokenbucket",
     "//pkg/ccl/multitenantccl/tenantcostserver/tenanttokenbucket:tenanttokenbucket_test",
     "//pkg/ccl/multitenantccl/tenantcostserver:tenantcostserver",
@@ -1814,6 +1812,8 @@ GO_TARGETS = [
     "//pkg/multitenant/tenantcapabilities:tenantcapabilities",
     "//pkg/multitenant/tenantcapabilities:tenantcapabilities_test",
     "//pkg/multitenant/tenantcapabilitiespb:tenantcapabilitiespb",
+    "//pkg/multitenant/tenantcostclient:tenantcostclient",
+    "//pkg/multitenant/tenantcostclient:tenantcostclient_test",
     "//pkg/multitenant/tenantcostmodel:tenantcostmodel",
     "//pkg/multitenant/tenantcostmodel:tenantcostmodel_test",
     "//pkg/multitenant:multitenant",
