+++ pkg/ccl/multitenantccl/doc.go
@@ -7,6 +7,6 @@ package multitenantccl
 
 import (
 	// Imports for the CCL init hooks.
-	_ "github.com/cockroachdb/cockroach/pkg/ccl/multitenantccl/tenantcostclient"
 	_ "github.com/cockroachdb/cockroach/pkg/ccl/multitenantccl/tenantcostserver"
+	_ "github.com/cockroachdb/cockroach/pkg/multitenant/tenantcostclient"
 )
